# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateQrResponse:
    properties:
      code:
        type: string
      expires_at:
        type: string
      id:
        type: string
      qr_code:
        items:
          type: integer
        type: array
    type: object
...
  model.Attendance:
    properties:
      attendanceId:
        type: string
      checkIn:
        type: string
      checkOut:
        type: string
      createdAt:
        type: string
      date:
        type: string
      id:
        type: integer
      location:
        type: string
      status:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.QRCode:
    properties:
      code:
        type: string
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: integer
      isUsed:
        type: boolean
      qrcodeId:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  model.Role:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      roleId:
        type: string
      updatedAt:
        type: string
    type: object
  model.Session:
    properties:
      createdAt:
        type: string
      expiresAt:
        type: string
      id:
        type: integer
      ipaddress:
        type: string
      isRevoked:
        type: boolean
      refreshToken:
        type: string
      sessionId:
        type: string
      updatedAt:
        type: string
      userAgent:
        type: string
      users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.User:
    properties:
      attendances:
        items:
          $ref: '#/definitions/model.Attendance'
        type: array
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      qrcode:
        $ref: '#/definitions/model.QRCode'
      roles:
        items:
          $ref: '#/definitions/model.Role'
        type: array
      sessions:
        items:
          $ref: '#/definitions/model.Session'
        type: array
      updatedAt:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
  types.ErrorResponse:
    properties:
      error: {}
      message:
        example: Unauthorized
        type: string
      status:
        example: 401
        type: integer
    type: object
  types.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:5000
info:
  contact:
    email: ilham15suandi@gmail.com
    name: Ilham Suandi
    url: https://github.com/ilhamSuandi
  description: Growy API
  title: Growy API
  version: 1.0.0
paths:
...
  /qrcode/{userId}:
    get:
      consumes:
      - application/json
      description: daily checkin user
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully created qrcode
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateQrResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: error getting qrcode
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: check in user
      tags:
      - QrCode
    post:
      consumes:
      - application/json
      description: manually re-generating user qrcode
      parameters:
      - description: userId
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: successfully created qrcode
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateQrResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: error creating qrcode
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: re-generate user qrcode
      tags:
      - QrCode
  /users:
    get:
      consumes:
      - application/json
      description: getting all users information
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get all users
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Users
      tags:
      - User
  /users/{userID}:
    get:
      consumes:
      - application/json
      description: getting specific user information
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get all users
          schema:
            allOf:
            - $ref: '#/definitions/types.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get User by UserId
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: 'Example Value: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"